+++ apps/xeml/src/cli/commands.js
@@ -2,6 +2,7 @@ exports.Commands = {
     'clean': 'Remove auto-generated files.',
     'list-modifiers': 'List built-in modifiers.',
     'build': 'Generate database scripts and entity models.',
+    'build-api': 'Generate API source code.',
     'build-api-client': 'Generate API client-side code.',
     'migrate': 'Create database structure.',
     'import': 'Import data set.',
@@ -24,6 +25,7 @@ exports.getCommandOptions = (cli, command) => {
             break;
 
         case 'build':
+        case 'build-api':
         case 'build-api-client':
             break;