+++ packages/validators/src/makeValidator.js
@@ -1,3 +1,5 @@
+import { template } from '@kitmi/utils';
+
 /**
  * Wrap the validate function into a validator.
  * @param {Function} validateFunc - Validate function to return true or false
@@ -9,7 +11,10 @@ const makeValidator = (validateFunc, message, checkNull) => {
     const functor = (value, options, meta, context) => {
         const validated = validateFunc(value, options, context);
         if (!validated) {
-            return [false, context.i18n?.t ? context.i18n.t(message, { value, options }) : message];
+            return [
+                false,
+                context.i18n?.t ? context.i18n.t(message, { value, options }) : template(message, { value, options }),
+            ];
         }
 
         return [true];